requirements.txt
colorlog==5.0.1
numpy==1.26.4
Pillow==10.3.0
torch==2.2.2
torchvision==0.17.2
